made set celsius and set farenheit commands work with floating point numbers

+++ hottube.ino
@@ -60,10 +60,10 @@ void setup() {
 
 void sendResponse(EthernetClient* client) {
   if (strncmp("GET /sc/", (char*)buffer, 8) == 0) {
-    set_celsius = (unsigned long)atoi(buffer+8);
+    set_celsius = atof(buffer+8);
   }
   else if (strncmp("GET /sf/", (char*)buffer, 8) == 0) {
-    set_celsius = farenheitToCelsius((unsigned long)atoi(buffer+8));
+    set_celsius = farenheitToCelsius(atof(buffer+8));
   }
   else if (strncmp("GET /j/off", (char*)buffer, 10) == 0) {
     // deactivate jets
